+++ src/lorhammer/main.go
@@ -9,6 +9,7 @@ import (
 	"lorhammer/src/tools"
 	"net/http"
 	"runtime"
+	"time"
 
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/sirupsen/logrus"
@@ -30,6 +31,7 @@ func main() {
 	nsAddress := flag.String("ns-address", "127.0.0.1:1700", "NetworkServer ip:port address")
 	logInfo := flag.Bool("vv", false, "log infos")
 	logDebug := flag.Bool("vvv", false, "log debugs")
+	maxWaitOrchestratorTime := flag.Duration("max-wait-orchestrator", 1*time.Minute, "The maximum time to wait for first communication with orchestrator")
 	flag.Parse()
 
 	if *showVersion {
@@ -90,7 +92,7 @@ func main() {
 		}
 
 		// LINK TO ORCHESTRATOR
-		lorhammerAddedChan := command.Start(mqttClient, hostname)
+		lorhammerAddedChan := command.Start(mqttClient, hostname, *maxWaitOrchestratorTime)
 		listenMqtt(mqttClient, []string{tools.MqttLorhammerTopic, tools.MqttLorhammerTopic + "/" + hostname}, hostname, lorhammerAddedChan, prometheus)
 	}