+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * **BREAKING** remove 'sleepAfterDeployTime' in deploy
   * **BREAKING** add 'requieredLorhammer' and 'maxWaitLorhammerTime' in scenario file
   * **BREAKING** remove ramp tytpe test because too complexe to manage in multiple scenarios
+  * add a flag to lorhammer **--max-wait-orchestrator** to ask for an orchestrator for a max period (1 minute by default)
 * **CONSUL** remove consul and pass mqtt url directly, permit to deploy more easily and to launch multiple orchestrator [issues/66](https://gitlab.com/itk.fr/lorhammer/issues/66)
 * **LORHAMMER** allow to set port used to be scrapped by prometheus [issues/64](https://gitlab.com/itk.fr/lorhammer/issues/64)
 * **ORCHESTRATOR** exit(1) if an error occur during scenario (previously exit > 0 only when checker fail) [issues/67](https://gitlab.com/itk.fr/lorhammer/issues/67) + repair bug on ramp type test